+++ packages/vsce/src/trees/CICSRegionTree.ts
@@ -24,9 +24,9 @@ import {
   TransactionMeta,
   URIMapMeta,
   WebServiceMeta,
-  JVMServerMeta
+  JVMServerMeta,
+  RemoteFileMeta
 } from "../doc";
-import { ResourceContainer } from "../resources";
 import { getIconByStatus } from "../utils/iconUtils";
 import { CICSPlexTree } from "./CICSPlexTree";
 import { CICSResourceContainerNode } from "./CICSResourceContainerNode";
@@ -75,7 +75,7 @@ export class CICSRegionTree extends CICSTreeNode implements ICICSTreeNode {
         this.children.push(this.buildResourceContainerNode([TransactionMeta]));
       }
       if (config.get<boolean>("LocalFile", true)) {
-        this.children.push(this.buildResourceContainerNode([LocalFileMeta]));
+        this.children.push(this.buildResourceContainerNode([LocalFileMeta, RemoteFileMeta], "Files"));
       }
       if (config.get<boolean>("Task", true)) {
         this.children.push(this.buildResourceContainerNode([TaskMeta]));